refactor(tools): split pipeline_report_result.go into focused sub-files (Phase 3a)
Extract verdict parsing, phase transition logic, and artifact guards into
dedicated files: verdict_parser.go, phase_transition.go, artifact_guard.go.
pipeline_report_result.go now retains only the core types and entry points.
Add verdict_parser_test.go with direct unit tests for determineTransition
and handlePhase6Transition.
