+++ mcp-server/internal/tools/pipeline_report_result.go
@@ -5,11 +5,7 @@ package tools
 import (
 	"context"
 	"fmt"
-	"os"
-	"path/filepath"
-	"sort"
 	"strings"
-	"time"
 
 	"github.com/mark3labs/mcp-go/mcp"
 	"github.com/mark3labs/mcp-go/server"
@@ -19,30 +15,6 @@ import (
 	"github.com/hiromaily/claude-forge/mcp-server/internal/state"
 	"github.com/hiromaily/claude-forge/mcp-server/internal/validation"
 )
-
-// phaseRevType maps review phases to their revision type passed to RevisionBump.
-//
-//nolint:gochecknoglobals // package-level lookup table for phase revision types
-var phaseRevType = map[string]string{
-	"phase-3b": "design",
-	"phase-4b": "tasks",
-}
-
-// reviewArtifactFile maps review phases to the artifact file that contains the verdict.
-//
-//nolint:gochecknoglobals // package-level lookup table for review artifact files
-var reviewArtifactFile = map[string]string{
-	"phase-3b": "review-design.md",
-	"phase-4b": "review-tasks.md",
-}
-
-// phaseAgentName maps review phases to the agent name used for pattern accumulation.
-//
-//nolint:gochecknoglobals // package-level lookup table for phase agent names
-var phaseAgentName = map[string]string{
-	"phase-3b": "design-reviewer",
-	"phase-4b": "task-reviewer",
-}
 
 // reportResultOutcome is the typed result of the report-result core logic.
 // It is returned by reportResultCore and consumed by both PipelineReportResultHandler
@@ -175,575 +147,4 @@ func handleReportResult(sm *state.StateManager, kb *history.KnowledgeBase, in re
 		return errorf("%v", err)
 	}
 	return okJSON(out)
-}
-
-// determineTransition decides the correct state transition and returns a partial response.
-//
-//nolint:gocyclo // complexity is inherent in the dispatch table
-func determineTransition(
-	sm *state.StateManager,
-	kb *history.KnowledgeBase,
-	in reportResultInput,
-	results []validation.ArtifactResult,
-	artifactWritten string,
-	warnings *[]string,
-) (reportResultOutcome, error) {
-	// Step 7: Review phases (phase-3b, phase-4b) — parse verdict and decide.
-	if revType, ok := phaseRevType[in.phase]; ok {
-		artifactFile, knownFile := reviewArtifactFile[in.phase]
-		if !knownFile {
-			// Fallback: complete the phase without verdict parsing.
-			if err := sm.PhaseComplete(in.workspace, in.phase); err != nil {
-				return reportResultOutcome{}, err
-			}
-			return reportResultOutcome{
-				StateUpdated:    true,
-				ArtifactWritten: artifactWritten,
-				NextActionHint:  "proceed",
-			}, nil
-		}
-
-		verdict, findings, err := orchestrator.ParseVerdict(filepath.Join(in.workspace, artifactFile))
-		if err != nil {
-			return reportResultOutcome{}, err
-		}
-
-		findings = nonNilSlice(findings)
-
-		// Accumulate review findings into the pattern knowledge base (fail-open).
-		agentName := phaseAgentName[in.phase]
-		if accumErr := kb.Patterns.Accumulate(findings, agentName, time.Now().UTC()); accumErr != nil {
-			*warnings = append(*warnings, "pattern accumulation warning: "+accumErr.Error())
-		}
-
-		switch verdict {
-		case orchestrator.VerdictRevise:
-			if err := sm.RevisionBump(in.workspace, revType); err != nil {
-				return reportResultOutcome{}, err
-			}
-			return reportResultOutcome{
-				StateUpdated:    true,
-				ArtifactWritten: artifactWritten,
-				VerdictParsed:   string(verdict),
-				Findings:        findings,
-				NextActionHint:  "revision_required",
-			}, nil
-		default:
-			// APPROVE, APPROVE_WITH_NOTES, or UNKNOWN — all advance the phase.
-			if err := sm.PhaseComplete(in.workspace, in.phase); err != nil {
-				return reportResultOutcome{}, err
-			}
-			return reportResultOutcome{
-				StateUpdated:    true,
-				ArtifactWritten: artifactWritten,
-				VerdictParsed:   string(verdict),
-				Findings:        findings,
-				NextActionHint:  "proceed",
-			}, nil
-		}
-	}
-
-	// Step 8: Phase-6 — parse verdict from each impl-*.md.
-	if in.phase == "phase-6" {
-		return handlePhase6Transition(sm, in, results, artifactWritten)
-	}
-
-	// Step 9: All other phases — advance unless setup_only.
-	if in.setupOnly {
-		return reportResultOutcome{
-			StateUpdated:    true,
-			ArtifactWritten: artifactWritten,
-			NextActionHint:  "setup_continue",
-		}, nil
-	}
-
-	// Step 9b: Phase-5 special handling — do not advance if pending tasks remain.
-	// After a parallel batch completes, there may be sequential tasks still pending.
-	// Re-enter handlePhaseFive by returning "setup_continue" instead of advancing.
-	if in.phase == "phase-5" {
-		// Auto-mark tasks as completed when their impl-N.md artifact exists.
-		// The implementer agent writes impl-N.md but may not call task_update
-		// explicitly, so we reconcile task status from artifact presence.
-		// Batch all updates in a single transaction to avoid O(N) disk I/O.
-		// Also detect parallel batch completion and set NeedsBatchCommit flag.
-		if updateErr := sm.Update(func(st *state.State) error {
-			newlyCompleted := 0
-			for k, t := range st.Tasks {
-				if t.ImplStatus == "completed" {
-					continue
-				}
-				implFile := filepath.Join(in.workspace, "impl-"+k+".md")
-				if _, statErr := os.Stat(implFile); statErr == nil {
-					if t.ExecutionMode == "parallel" {
-						newlyCompleted++
-					}
-					t.ImplStatus = "completed"
-					st.Tasks[k] = t
-				}
-			}
-			// If parallel tasks were just completed, signal batch commit needed.
-			if newlyCompleted > 0 {
-				st.NeedsBatchCommit = true
-			}
-			return nil
-		}); updateErr != nil {
-			return reportResultOutcome{}, updateErr
-		}
-
-		// Re-read state after potential updates.
-		s, err := sm.GetState()
-		if err != nil {
-			return reportResultOutcome{}, err
-		}
-		hasPending := false
-		for _, t := range s.Tasks {
-			if t.ImplStatus != "completed" {
-				hasPending = true
-				break
-			}
-		}
-		// Also hold in phase-5 if a batch commit is pending (e.g. last parallel
-		// batch just completed — all tasks done but git commit not yet run).
-		if hasPending || s.NeedsBatchCommit {
-			return reportResultOutcome{
-				StateUpdated:    true,
-				ArtifactWritten: artifactWritten,
-				NextActionHint:  "setup_continue",
-			}, nil
-		}
-
-		// Phase 5 completion gate: verify impl file count matches task count.
-		// This is a deterministic safety check — even if task status says all
-		// completed, the actual impl-{N}.md files must exist for every task.
-		// When missing files are found, reset ImplStatus so the engine
-		// re-dispatches implementers on the next pipeline_next_action call.
-		if missing := missingImplFiles(in.workspace, s.Tasks); len(missing) > 0 {
-			if updateErr := sm.Update(func(st *state.State) error {
-				for _, k := range missing {
-					if t, ok := st.Tasks[k]; ok {
-						t.ImplStatus = ""
-						st.Tasks[k] = t
-					}
-				}
-				return nil
-			}); updateErr != nil {
-				return reportResultOutcome{}, updateErr
-			}
-			return reportResultOutcome{
-				StateUpdated:    true,
-				ArtifactWritten: artifactWritten,
-				NextActionHint:  "setup_continue",
-				Warning:         "phase-5 completion blocked: missing impl files for tasks: " + strings.Join(missing, ", ") + " — ImplStatus reset to pending",
-			}, nil
-		}
-
-		// All tasks complete — clear any completed_fail retry state so the
-		// engine dispatches fresh reviewers after the retry implementer ran.
-		if err := clearCompletedFailTasks(sm, in.workspace); err != nil {
-			return reportResultOutcome{}, err
-		}
-	}
-
-	// Phase-4 (task-decomposer) completion gate: apply deterministic workflow
-	// rules from .specs/instructions.md. If violations exist, write
-	// review-tasks.md and emit revision_required so the engine re-dispatches
-	// task-decomposer with the findings.
-	if in.phase == "phase-4" {
-		if resp, handled, err := applyWorkflowRules(sm, in.workspace, artifactWritten); err != nil {
-			return reportResultOutcome{}, err
-		} else if handled {
-			return resp, nil
-		}
-	}
-
-	if err := sm.PhaseComplete(in.workspace, in.phase); err != nil {
-		return reportResultOutcome{}, err
-	}
-	return reportResultOutcome{
-		StateUpdated:    true,
-		ArtifactWritten: artifactWritten,
-		NextActionHint:  "proceed",
-	}, nil
-}
-
-// handlePhase6Transition processes phase-6 results, parsing verdicts from review-*.md
-// files and deterministically updating task ReviewStatus in state.json.
-//
-// Key invariant: after this function returns, every task whose review-{k}.md
-// contains a PASS/PASS_WITH_NOTES verdict will have its ReviewStatus set in
-// state.json. This prevents the engine from re-dispatching reviewers for
-// already-reviewed tasks.
-//
-//nolint:gocyclo // complexity is inherent in multi-task verdict reconciliation
-func handlePhase6Transition(
-	sm *state.StateManager,
-	in reportResultInput,
-	results []validation.ArtifactResult,
-	artifactWritten string,
-) (reportResultOutcome, error) {
-	allFindings := []orchestrator.Finding{}
-	var verdictParsed string
-	anyFail := false
-
-	for _, result := range results {
-		if result.File == "" {
-			continue
-		}
-
-		verdict, findings, err := orchestrator.ParseVerdict(filepath.Join(in.workspace, result.File))
-		if err != nil {
-			// File I/O error — treat as fail.
-			anyFail = true
-			continue
-		}
-
-		if findings != nil {
-			allFindings = append(allFindings, findings...)
-		}
-
-		if verdictParsed == "" {
-			verdictParsed = string(verdict)
-		}
-
-		// Only VerdictFail triggers retry; VerdictPassWithNotes is treated as passing.
-		if verdict == orchestrator.VerdictFail {
-			anyFail = true
-		}
-	}
-
-	if anyFail {
-		// Deterministic: record the FAIL verdict in state so the engine can
-		// dispatch the retry via ReviewStatus without re-reading the review file.
-		// This also prevents double-increment if pipeline_next_action is called
-		// multiple times before the orchestrator calls pipeline_report_result.
-		if updateErr := sm.Update(func(st *state.State) error {
-			for _, result := range results {
-				if result.File == "" || result.VerdictFound != state.VerdictFail {
-					continue
-				}
-				taskKey := reviewFileTaskKey(result.File)
-				if taskKey == "" {
-					continue
-				}
-				if t, ok := st.Tasks[taskKey]; ok {
-					t.ImplRetries++
-					t.ReviewStatus = state.TaskStatusCompletedFail
-					st.Tasks[taskKey] = t
-				}
-			}
-			return nil
-		}); updateErr != nil {
-			return reportResultOutcome{}, updateErr
-		}
-		return reportResultOutcome{
-			StateUpdated:    true,
-			ArtifactWritten: artifactWritten,
-			VerdictParsed:   verdictParsed,
-			Findings:        allFindings,
-			NextActionHint:  "retry_impl",
-		}, nil
-	}
-
-	// Deterministic: reconcile ReviewStatus from review file verdicts.
-	// This ensures the engine (handlePhaseSix) never re-processes a task
-	// whose review file already contains a passing verdict.
-	// Read verdicts outside the lock to avoid I/O inside a critical section.
-	s, err := sm.GetState()
-	if err != nil {
-		return reportResultOutcome{}, err
-	}
-	type verdictUpdate struct {
-		key    string
-		status string
-	}
-	var verdictUpdates []verdictUpdate
-	for k, t := range s.Tasks {
-		if t.ImplStatus != state.TaskStatusCompleted {
-			continue
-		}
-		if t.ReviewStatus == state.TaskStatusCompletedPass ||
-			t.ReviewStatus == state.TaskStatusCompletedPassNote {
-			continue
-		}
-		reviewFile := filepath.Join(in.workspace, "review-"+k+".md")
-		if _, statErr := os.Stat(reviewFile); statErr != nil {
-			continue
-		}
-		verdict, _, vErr := orchestrator.ParseVerdict(reviewFile)
-		if vErr != nil {
-			continue
-		}
-		switch verdict { //nolint:exhaustive // only PASS variants update state; FAIL/REVISE/UNKNOWN are intentionally ignored here
-		case orchestrator.VerdictPass:
-			verdictUpdates = append(verdictUpdates, verdictUpdate{key: k, status: state.TaskStatusCompletedPass})
-		case orchestrator.VerdictPassWithNotes:
-			verdictUpdates = append(verdictUpdates, verdictUpdate{key: k, status: state.TaskStatusCompletedPassNote})
-		}
-	}
-	if len(verdictUpdates) > 0 {
-		if updateErr := sm.Update(func(st *state.State) error {
-			for _, u := range verdictUpdates {
-				t := st.Tasks[u.key]
-				t.ReviewStatus = u.status
-				st.Tasks[u.key] = t
-			}
-			return nil
-		}); updateErr != nil {
-			return reportResultOutcome{}, updateErr
-		}
-	}
-
-	// Check whether any task still needs a review.
-	s, err = sm.GetState()
-	if err != nil {
-		return reportResultOutcome{}, err
-	}
-	for _, t := range s.Tasks {
-		if t.ImplStatus != state.TaskStatusCompleted {
-			continue
-		}
-		if t.ReviewStatus != state.TaskStatusCompletedPass &&
-			t.ReviewStatus != state.TaskStatusCompletedPassNote {
-			// Task needs review — hold in phase-6.
-			return reportResultOutcome{
-				StateUpdated:    true,
-				ArtifactWritten: artifactWritten,
-				VerdictParsed:   verdictParsed,
-				Findings:        allFindings,
-				NextActionHint:  "setup_continue",
-			}, nil
-		}
-	}
-
-	// Phase 6 completion gate: verify review file count matches task count.
-	// When missing files are found, reset ReviewStatus so the engine
-	// re-dispatches reviewers on the next pipeline_next_action call.
-	if missing := missingReviewFiles(in.workspace, s.Tasks); len(missing) > 0 {
-		if updateErr := sm.Update(func(st *state.State) error {
-			for _, k := range missing {
-				if t, ok := st.Tasks[k]; ok {
-					t.ReviewStatus = ""
-					st.Tasks[k] = t
-				}
-			}
-			return nil
-		}); updateErr != nil {
-			return reportResultOutcome{}, updateErr
-		}
-		return reportResultOutcome{
-			StateUpdated:    true,
-			ArtifactWritten: artifactWritten,
-			VerdictParsed:   verdictParsed,
-			Findings:        allFindings,
-			NextActionHint:  "setup_continue",
-			Warning:         "phase-6 completion blocked: missing review files for tasks: " + strings.Join(missing, ", ") + " — ReviewStatus reset to pending",
-		}, nil
-	}
-
-	if err := sm.PhaseComplete(in.workspace, in.phase); err != nil {
-		return reportResultOutcome{}, err
-	}
-	return reportResultOutcome{
-		StateUpdated:    true,
-		ArtifactWritten: artifactWritten,
-		VerdictParsed:   verdictParsed,
-		Findings:        allFindings,
-		NextActionHint:  "proceed",
-	}, nil
-}
-
-// reviewFileTaskKey extracts the task key from a review file basename.
-// e.g. "review-1.md" → "1", "review-task-abc.md" → "task-abc".
-// Returns "" if the filename does not match the review-*.md or impl-*.md pattern.
-func reviewFileTaskKey(filename string) string {
-	base := filepath.Base(filename)
-	switch {
-	case strings.HasPrefix(base, "review-") && strings.HasSuffix(base, ".md"):
-		return strings.TrimSuffix(strings.TrimPrefix(base, "review-"), ".md")
-	case strings.HasPrefix(base, "impl-") && strings.HasSuffix(base, ".md"):
-		return strings.TrimSuffix(strings.TrimPrefix(base, "impl-"), ".md")
-	}
-	return ""
-}
-
-// missingArtifactFiles returns task keys whose {prefix}{key}.md file does not exist on disk.
-// Human-gate tasks are excluded because they produce no artifact file.
-func missingArtifactFiles(workspace, prefix string, tasks map[string]state.Task) []string {
-	var missing []string
-	for k, t := range tasks {
-		if t.ExecutionMode == state.ExecModeHumanGate {
-			continue
-		}
-		if _, err := os.Stat(filepath.Join(workspace, prefix+k+".md")); err != nil {
-			missing = append(missing, k)
-		}
-	}
-	sort.Strings(missing)
-	return missing
-}
-
-// missingImplFiles returns task keys whose impl-{key}.md file does not exist on disk.
-// Used as a deterministic completion gate for Phase 5 — prevents the phase from
-// advancing when some tasks lack implementation artifacts, regardless of task status.
-// Human-gate tasks are excluded because they are completed by user acknowledgement
-// and intentionally produce no impl file.
-func missingImplFiles(workspace string, tasks map[string]state.Task) []string {
-	return missingArtifactFiles(workspace, "impl-", tasks)
-}
-
-// missingReviewFiles returns task keys whose review-{key}.md file does not exist on disk.
-// Used as a deterministic completion gate for Phase 6 — prevents the phase from
-// advancing when some tasks lack review artifacts.
-// Human-gate tasks are excluded because they skip review and produce no review file.
-func missingReviewFiles(workspace string, tasks map[string]state.Task) []string {
-	return missingArtifactFiles(workspace, "review-", tasks)
-}
-
-// clearCompletedFailTasks resets ReviewStatus and removes stale review files for
-// tasks in the "completed_fail" state. Called from the phase-5 handler after a
-// retry implementer run so the engine dispatches a fresh reviewer on the next call.
-func clearCompletedFailTasks(sm *state.StateManager, workspace string) error {
-	return sm.Update(func(st *state.State) error {
-		for k, t := range st.Tasks {
-			if t.ReviewStatus != state.TaskStatusCompletedFail {
-				continue
-			}
-			// Delete the stale review file so the engine dispatches a fresh reviewer.
-			reviewFile := filepath.Join(workspace, "review-"+k+".md")
-			if err := os.Remove(reviewFile); err != nil && !os.IsNotExist(err) {
-				return err
-			}
-			t.ReviewStatus = ""
-			st.Tasks[k] = t
-		}
-		return nil
-	})
-}
-
-// applyWorkflowRules runs .specs/instructions.md rules against the phase-4
-// tasks.md output. Returns (handled=true, resp) when violations exist and
-// the caller should return without calling PhaseComplete. Returns
-// (handled=false, zero) in any of these pass-through cases:
-//   - tasks.md cannot be read or parsed (the existing artifact validator
-//     will surface the failure).
-//   - workspace is not under a .specs/ directory (flat layout — workflow
-//     rules do not apply).
-//   - no violations were found. This also covers the "no rules file"
-//     case: validation.LoadRules returns an empty rule set when
-//     .specs/instructions.md is absent, so Validate yields zero
-//     violations and we proceed unchanged. On the pass-through, any
-//     stale review-tasks.md left behind by a previous workflow-rules
-//     iteration is removed so the phase-4b task-reviewer can write a
-//     fresh file.
-//
-// Why here and not in validation.ValidateArtifacts: that API only checks
-// file presence and verdict tokens — it does not take the parsed tasks
-// map or the repo root. This helper owns the specific phase-4 wiring.
-//
-// On the violation path, this helper bumps state.Revisions.TaskRevisions via
-// sm.RevisionBump so the engine's retry-limit guard (handlePhaseFour) can
-// enforce MaxRevisionRetries on the next pipeline_next_action call. This
-// mirrors how VerdictRevise on phase-4b increments the same counter in
-// determineTransition.
-func applyWorkflowRules(sm *state.StateManager, workspace, artifactWritten string) (reportResultOutcome, bool, error) {
-	tasks, rules, reviewPath, ok, err := loadPhase4Context(workspace)
-	if !ok || err != nil {
-		return reportResultOutcome{}, false, err
-	}
-
-	violations := validation.Validate(tasks, rules)
-	if len(violations) == 0 {
-		// Pass-through: ensure any stale review-tasks.md from an earlier
-		// workflow-rules iteration in the same pipeline is removed so the
-		// phase-4b task-reviewer (handlePhaseFourB) writes a fresh file
-		// instead of reading a stale REVISE verdict and looping.
-		if err := os.Remove(reviewPath); err != nil && !os.IsNotExist(err) {
-			return reportResultOutcome{}, false, fmt.Errorf("remove stale %s: %w", state.ArtifactReviewTasks, err)
-		}
-		return reportResultOutcome{}, false, nil
-	}
-
-	resp, err := writeViolationResponse(sm, workspace, reviewPath, artifactWritten, violations)
-	if err != nil {
-		return reportResultOutcome{}, false, err
-	}
-	return resp, true, nil
-}
-
-// loadPhase4Context reads tasks.md, checks the workspace layout, and loads
-// workflow rules. Returns ok=false (no error) for silent pass-through cases
-// where workflow rules should not be applied.
-func loadPhase4Context(workspace string) (map[string]state.Task, *validation.WorkflowRules, string, bool, error) {
-	tasksData, err := os.ReadFile(filepath.Join(workspace, state.ArtifactTasks))
-	if err != nil {
-		// tasks.md missing: let the normal artifact validator handle it.
-		return nil, nil, "", false, nil
-	}
-
-	tasks, err := ParseTasksMd(string(tasksData))
-	if err != nil {
-		// Let the caller fail via artifact validation or parser errors.
-		return nil, nil, "", false, nil
-	}
-
-	// Workflow rules only apply when the workspace follows the
-	// .specs/<spec>/ layout. Any other layout (e.g. flat test fixtures)
-	// falls through silently — fail-open rather than mis-resolving the
-	// repo root into an unrelated directory.
-	if filepath.Base(filepath.Dir(workspace)) != ".specs" {
-		return nil, nil, "", false, nil
-	}
-	// Repo root is two levels up from a .specs/<spec-name>/ workspace.
-	repoRoot := filepath.Dir(filepath.Dir(workspace))
-	rules, err := validation.LoadRules(repoRoot)
-	if err != nil {
-		// Rule file exists but is malformed. Surface as an error so the
-		// user sees the parse failure instead of silently skipping.
-		return nil, nil, "", false, fmt.Errorf("load workflow rules: %w", err)
-	}
-
-	reviewPath := filepath.Join(workspace, state.ArtifactReviewTasks)
-	return tasks, rules, reviewPath, true, nil
-}
-
-// writeViolationResponse writes review-tasks.md, bumps TaskRevisions, and
-// builds the reportResultOutcome for a workflow-rules violation.
-//
-// Order matters: write review-tasks.md FIRST, then bump TaskRevisions.
-// If RevisionBump fails, the orchestrator still sees review-tasks.md on
-// the next pipeline_next_action call and handlePhaseFour re-dispatches
-// task-decomposer — we just lose one retry-limit tick, which is
-// recoverable. Reversing the order would risk bumping the counter with
-// no findings file on disk, which would waste a retry slot silently.
-func writeViolationResponse(sm *state.StateManager, workspace, reviewPath, artifactWritten string, violations []validation.Violation) (reportResultOutcome, error) {
-	if err := os.WriteFile(reviewPath, []byte(validation.FormatReviewFindings(violations)), 0o600); err != nil {
-		return reportResultOutcome{}, fmt.Errorf("write %s: %w", state.ArtifactReviewTasks, err)
-	}
-
-	// Bump TaskRevisions so handlePhaseFour enforces MaxRevisionRetries on
-	// the next pipeline_next_action call. This mirrors how determineTransition
-	// bumps the same counter when phase-4b parses a REVISE verdict — keeping
-	// the retry-limit enforcement in one place (the engine).
-	if err := sm.RevisionBump(workspace, state.RevTypeTasks); err != nil {
-		return reportResultOutcome{}, fmt.Errorf("revision bump (tasks): %w", err)
-	}
-
-	findings := make([]orchestrator.Finding, 0, len(violations))
-	for _, v := range violations {
-		findings = append(findings, orchestrator.Finding{
-			Severity: orchestrator.SeverityCritical,
-			Description: fmt.Sprintf("task %s (%s) violates rule %q: %s",
-				v.TaskKey, v.TaskTitle, v.RuleID, v.Reason),
-		})
-	}
-
-	return reportResultOutcome{
-		StateUpdated:    true,
-		ArtifactWritten: artifactWritten,
-		VerdictParsed:   "REVISE",
-		Findings:        findings,
-		NextActionHint:  "revision_required",
-		Warning: fmt.Sprintf("phase-4 workflow rules: %d violation(s) — see %s",
-			len(violations), state.ArtifactReviewTasks),
-	}, nil
 }
